// src/store/actions/users/interfaces.ts
export enum getUsersType {
  GET_USERS = 'GET_USERS',
  GET_USERS_FAIL = 'GET_USERS_FAIL',
  GET_USERS_SUCCESS = 'GET_USERS_SUCCESS',
  SHOW_USER = 'SHOW_USER',
  RESET_EDITING = 'RESET_EDITING',
  SHOW_WARNING = 'SHOW_WARNING',
  CANCEL_DELETE = 'CANCEL_DELETE',
  EDIT_USER_FAIL = 'EDIT_USER_FAIL',
  EDIT_USER_SUCCESS = 'EDIT_USER_SUCCESS'
}